Add, Update SV

+++ ATBM_PhanHe1/DAO/RegisterDAO.cs
@@ -47,10 +47,10 @@ namespace ATBM_PhanHe1.DAO
                 result.Add(register);
             }
             return result;
-
+        }
         public bool DeleteRegisterByID(string id)
         {
-            string query = string.Format("begin delete from tb_nhansu where MANV = {0}; end;", id);
+            string query = string.Format("begin delete from admin.tb_nhansu where MANV = {0}; end;", id);
             int result = DataProvider.Instance.ExecuteNonQuery(query);
             return result > 0;
             
@@ -58,7 +58,7 @@ namespace ATBM_PhanHe1.DAO
        
         public PersonelDTO GetPersonelByID(string personelID)
         {
-            string query = string.Format("select * from tb_nhansu where lower(MANV) like lower('%{0}%')", personelID);
+            string query = string.Format("select * from admin.tb_nhansu where lower(MANV) like lower('%{0}%')", personelID);
             DataTable data = DataProvider.Instance.ExecuteQuery(query);
             PersonelDTO result = new PersonelDTO(data.Rows[0]);
             return result;
